Complete RestClient and clean some repo
