+++ app/src/main/java/com/android3/siegertpclient/data/user/usersource/userRemote/UserRemoteDataSource.kt
@@ -4,82 +4,60 @@ import com.android3.siegertpclient.data.invitation.Invitation
 import com.android3.siegertpclient.data.team.teamsource.teamLocal.Team
 import com.android3.siegertpclient.data.tournament.Tournament
 import com.android3.siegertpclient.data.user.User
-import com.android3.siegertpclient.utils.RestClient
+import com.android3.siegertpclient.utils.RestClient2
 import com.android3.siegertpclient.utils.TokenUtil
 import retrofit2.Response
 
-class UserRemoteDataSource(private val userService: UserService) {
+class UserRemoteDataSource2 {
 
-    private val runtimeError = "Response of Server was not successful"
-
-    fun createNewUser (username: String, surname: String, firstName: String, userId : String) : User? {
+    suspend fun createNewUser(
+        username: String,
+        surname: String,
+        forename: String,
+        userId: String,
+        token: String
+    ): Response<User> {
         val user = hashMapOf<String, String>()
-        user["surname"] = surname
-        user["forename"] = firstName
         user["username"] = username
+        user["surname"] = surname
+        user["forename"] = forename
         user["userId"] = userId
-
-        val userCall = userService.createNewUser(user, TokenUtil.getBearerToken())
-        if (userCall.isCanceled) {
-            throw RuntimeException(runtimeError)
-        }
-        val response = userCall.execute()
-        return response.body()
+        return RestClient2.userService.createNewUser(user, token)
     }
 
-    fun getUserById (userId : String) : User? {
-        val userCall = userService.getUserById(userId, TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return response.body()
+    suspend fun getUserById(userId: String, token: String): Response<User> {
+        return RestClient2.userService.getUserById(userId, token)
     }
 
-    fun getUserByUsername (username : String) : User? {
-        val userCall = userService.getUserByUsername(username, TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return response.body()
+    suspend fun getUserByUsername(username: String, token: String): Response<User> {
+        return RestClient2.userService.getUserByUsername(username, token)
     }
 
-    fun getUsersTournaments (username: String) : List<Tournament>? {
-        val userCall = userService.getUsersTournaments(username, TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return response.body()
+    suspend fun getUsersTournaments(username: String, token: String): Response<List<Tournament>> {
+        return RestClient2.userService.getUsersTournaments(username, token)
     }
 
-    fun getUsersTeams (username: String) : List<Team>? {
-        val userCall = userService.getUserTeams(username, TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return response.body()
+    suspend fun getUsersTeams(username: String, token: String): Response<List<Team>> {
+        return RestClient2.userService.getUserTeams(username, token)
     }
 
-    fun getUsersInvitations (username: String) : List<Invitation>? {
-        val userCall = userService.getUserInvitations(username, TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return response.body()
+    suspend fun getUsersInvitations(username: String, token: String): Response<List<Invitation>> {
+        return RestClient2.userService.getUserInvitations(username, token)
     }
 
-    fun updateUserDetail (oldUsername: String, newUsername : String, forename : String, surname : String) : User? {
-        val userCall = userService.updateUserDetails(oldUsername,newUsername,forename,surname,
-            TokenUtil.getBearerToken())
-        val response = userCall.execute()
-        val uResp =  response.body()
-        if (!response.isSuccessful) {
-            throw RuntimeException(runtimeError)
-        }
-        return uResp
+    suspend fun updateUserDetail(
+        oldUsername: String,
+        newUsername: String,
+        newForename: String,
+        newSurname: String,
+        token: String
+    ): Response<User> {
+        return RestClient2.userService.updateUserDetails(
+            oldUsername,
+            newUsername,
+            newForename,
+            newSurname,
+            token
+        )
     }
 }
