+++ app/src/main/java/com/android3/siegertpclient/utils/RestClient2.kt
@@ -1,7 +1,10 @@
 package com.android3.siegertpclient.utils
 
+import com.android3.siegertpclient.data.game.gamesource.gameRemote.GameService
+import com.android3.siegertpclient.data.invitation.invitationsource.invitationRemote.InvitationService
+import com.android3.siegertpclient.data.team.teamsource.teamRemote.TeamService
 import com.android3.siegertpclient.data.tournament.tournamentsource.tournamentRemote.TournamentService2
-import com.android3.siegertpclient.data.user.usersource.userRemote.UserService2
+import com.android3.siegertpclient.data.user.usersource.userRemote.UserService
 import com.android3.siegertpclient.ui.dummyretrofit.util.Constants.Companion.BASE_URL
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -14,10 +17,21 @@ object RestClient2 {
             .build()
     }
 
-    val userService: UserService2 by lazy {
-        retrofit.create(UserService2::class.java)
+    val userService: UserService by lazy {
+        retrofit.create(UserService::class.java)
     }
 
+    val invitationService: InvitationService by lazy {
+        retrofit.create(InvitationService::class.java)
+    }
+
+    val teamService: TeamService by lazy {
+        retrofit.create(TeamService::class.java)
+    }
+
+    val gameService: GameService by lazy {
+        retrofit.create(GameService::class.java)
+    }
 
     val tournamentService: TournamentService2 by lazy {
         retrofit.create(TournamentService2::class.java)
